+++ packages/service-fetchq/src/index.js
@@ -107,7 +107,7 @@ module.exports = ({ registerAction, registerHook }) => {
     hook: '$FASTIFY_TDD_ROUTE?',
     name: hooks.SERVICE_NAME,
     trace: __filename,
-    handler: ({ registerTddRoute }) => {
+    handler: ({ registerTddRoute }, { createHook }) => {
       const schemaFields = {
         type: 'object',
         properties: {
@@ -133,6 +133,118 @@ module.exports = ({ registerAction, registerHook }) => {
         handler: (request) => {
           const { q: sql } = request.body;
           return request.fetchq.pool.query(sql);
+        },
+      });
+
+      registerTddRoute({
+        method: 'GET',
+        url: '/fetchq/:queue/:subject',
+        schema: {
+          params: {
+            type: 'object',
+            properties: {
+              queue: { type: 'string' },
+              subject: { type: 'string' },
+            },
+            required: ['queue', 'subject'],
+          },
+        },
+        handler: async (request, reply) => {
+          const { queue, subject } = request.params;
+          const result = await request.fetchq.pool.query(
+            `
+            SELECT * FROM "fetchq_data"."${queue}__docs"
+            WHERE "subject" = '${subject}'
+            `,
+          );
+
+          if (!result.rowCount) {
+            reply.status(404).send('Document not found');
+            return;
+          }
+
+          return result.rows[0];
+        },
+      });
+
+      registerTddRoute({
+        method: 'POST',
+        url: '/fetchq/await/:queue/:subject',
+        schema: {
+          params: {
+            type: 'object',
+            properties: {
+              queue: { type: 'string' },
+              subject: { type: 'string' },
+            },
+            required: ['queue', 'subject'],
+          },
+          body: {
+            type: 'object',
+            properties: {
+              status: { type: 'number', default: 3 },
+              timeout: { type: 'number', default: 1000 },
+              interval: { type: 'number', default: 100 },
+            },
+          },
+        },
+        handler: async (request, reply) => {
+          const { queue, subject } = request.params;
+          const { status, timeout, interval } = request.body;
+
+          let checkInterval = null;
+          let checkTimeout = null;
+          let doc = null;
+
+          checkInterval = setInterval(async () => {
+            const result = await request.fetchq.pool.query(
+              `
+              SELECT * FROM "fetchq_data"."${queue}__docs"
+              WHERE "subject" = '${subject}'
+              `,
+            );
+
+            // Eagerly exit the processin in case there is no document at all
+            if (!result.rowCount) {
+              return;
+            }
+
+            // Persist last retrieved document version
+            doc = result.rows[0];
+
+            // Verify that the condition is met and break the loop
+            if (Number(doc.status) === status) {
+              clearTimeout(checkTimeout);
+              clearInterval(checkInterval);
+              reply.status(200).send(doc);
+            }
+          }, interval);
+
+          // Interrupt the execution at the provided timeout
+          checkTimeout = setTimeout(() => {
+            clearInterval(checkInterval);
+            reply.status(412).send(doc);
+          }, timeout);
+        },
+      });
+
+      registerTddRoute({
+        method: 'GET',
+        url: '/fetchq/state/reset',
+        handler: async (request) => {
+          const { fetchq } = request;
+          const query = fetchq.pool.query.bind(fetchq.pool);
+
+          // Destroy and recreate Fetchq's schema and related
+          // data structure and workers
+          await fetchq.stop();
+          await query(
+            'DROP SCHEMA "fetchq_data" CASCADE; DROP SCHEMA "fetchq" CASCADE;',
+          );
+          await fetchq.boot();
+
+          await createHook.serie(hooks.FETCHQ_TDD_STATE_RESET, { query });
+          return '+ok';
         },
       });
     },